+++ src/main/java/com/application/close/links/controller/ModMqttLinksController.java
@@ -19,23 +19,21 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ModMqttLinksController {
 
-	private ModMqttLinksService linksService;
+	private final ModMqttLinksService linksService;
 
 	@Operation(summary = "Link Modbus TCP with MQTT Param", description = "Creates a mapping between a Modbus TCP client and an MQTT parameter.")
 	@PostMapping("/connect/tcpId/{tcpId}/paramId/{paramId}")
 	public ResponseEntity<String> connectModToMqtt(@PathVariable int tcpId, @PathVariable int paramId) {
 
 		linksService.connectModToMqtt(tcpId, paramId);
-		return ResponseEntity
-				.ok("Linked Modbus TCP ID " + tcpId + " with MQTT Param ID " + paramId + " successfully.");
+		return ResponseEntity.ok("Linked Modbus TCP ID " + tcpId + " with MQTT Param ID " + paramId + " successfully.");
 	}
 
 	@Operation(summary = "Unlink Modbus TCP from MQTT Param", description = "Removes the existing mapping between a Modbus TCP client and an MQTT parameter.")
 	@DeleteMapping("/disconnect/tcpId/{tcpId}")
 	public ResponseEntity<String> disconnectModFromMqtt(@PathVariable int tcpId) {
 		linksService.disconnectModFromMqtt(tcpId);
-		return ResponseEntity
-				.ok("Unlinked Modbus TCP ID " + tcpId + " from its associated MQTT Param successfully.");
+		return ResponseEntity.ok("Unlinked Modbus TCP ID " + tcpId + " from its associated MQTT Param successfully.");
 	}
 
 }