+++ src/main/java/com/application/close/links/service/BridgeExecutorService.java
@@ -19,6 +19,8 @@ public interface BridgeExecutorService {
 
 	List<BridgeExecutor> getAll();
 
+	List<Integer> getAllTcpId();
+
 	void deleteById(int executerId);
 
 }